test: update fiat rates tests
