+++ src/index.test.ts
@@ -351,75 +351,4 @@ test('should be ble to get marketData', async () => {
     expect(wallet.state.marketData.mainnet?.length).toBeGreaterThan(10);
 });
 
-test('should be able validate fiatRates for all mainnet assets', async () => {
-    const wallet = await setupWallet(defaultWalletOptions);
-    await wallet.dispatch.createWallet({
-        key: '0x1234567890123456789012345678901234567890',
-        mnemonic: 'test',
-        imported: true,
-    });
-    await wallet.dispatch.unlockWallet({
-        key: '0x1234567890123456789012345678901234567890',
-    });
-    expect(wallet.state.wallets.length).toBe(1);
-    expect(wallet.state.wallets[0].imported).toBe(true);
-    expect(wallet.state.unlockedAt).not.toBe(0);
 
-    await wallet.dispatch.initializeAnalyticsPreferences({
-        accepted: true,
-    });
-
-    expect(wallet.state.analytics.userId).not.toBe(null);
-    expect(wallet.state.analytics.acceptedDate).not.toBe(0);
-    expect(wallet.state.analytics.askedDate).not.toBe(0);
-    expect(wallet.state.analytics.askedTimes).toBe(0);
-    expect(wallet.state.analytics.notAskAgain).toBe(false);
-
-    const walletId = wallet.state.activeWalletId;
-    const mainnetAccounts = wallet?.state?.enabledAssets?.mainnet?.[walletId];
-    expect(mainnetAccounts).not.toBeNull();
-
-    // update balance this will generate address
-    if (mainnetAccounts) {
-        await wallet.dispatch.updateBalances({
-            network: Network.Mainnet,
-            walletId: walletId,
-            assets: mainnetAccounts,
-        });
-    }
-
-    const account = wallet.state.accounts?.[walletId]?.mainnet?.[1];
-    expect(account?.chain).toBe(ChainId.Ethereum);
-    const ethAccountId = account?.id;
-    const ethereumAddress = account?.addresses[0];
-    expect(ethereumAddress).not.toBeNull();
-
-    // update account balances eth
-    if (ethAccountId && ethereumAddress) {
-        await wallet.dispatch.updateAccountBalance({
-            network: Network.Mainnet,
-            walletId: wallet.state.activeWalletId,
-            accountId: ethAccountId,
-        });
-
-        if (mainnetAccounts) {
-            await wallet.dispatch.updateBalances({
-                network: Network.Mainnet,
-                walletId: wallet.state.activeWalletId,
-                assets: mainnetAccounts,
-            });
-
-            // update fiat rates for all assets
-            await wallet.dispatch.updateFiatRates({
-                assets: mainnetAccounts
-            });
-        }
-    }
-    // validate fiat rates & validate balances
-    expect((Object.keys(wallet.state.fiatRates)).length).toBeGreaterThan(0)
-    const fiatRatesObject = Object.values(wallet.state.fiatRates);
-    for (let i = 0; i < fiatRatesObject.length; i++) {
-        const fiatRate = fiatRatesObject[i];
-        expect(fiatRate).toBeGreaterThan(0);
-    }
-})
